+++ android/src/main/kotlin/io/adrop/adrop_ads/native/FlutterAdropNativeAd.kt
@@ -57,6 +57,7 @@ class FlutterAdropNativeAd(
     }
 
     override fun destroy() {
+        nativeAd.destroy()
     }
 
     override fun onAdClicked(ad: AdropNativeAd) {